feat: unify webrtc gateway
